Split behavior simulator modules
