+++ crates/bevyhavior_simulator/src/config.rs
@@ -1,0 +1,54 @@
+use std::{net::SocketAddr, time::Duration};
+
+use color_eyre::{Result, eyre::eyre};
+use types::parameters::BehaviorParameters;
+
+pub const DEFAULT_TICK_DURATION: Duration = Duration::from_millis(10);
+
+#[derive(bevy::prelude::Resource, Clone, Debug)]
+pub struct SimulationConfig {
+    pub walk_translation_speed: f32,
+    pub walk_rotation_speed: f32,
+    pub walk_with_velocity_scale: f32,
+    pub kick_ball_speed_rumpelstilzchen: f32,
+    pub kick_ball_speed_schlong: f32,
+    pub kick_cooldown: Duration,
+    pub ball_friction_per_second: f32,
+    pub ball_visibility_range: f32,
+    pub ball_visibility_angle: f32,
+    pub robot_radius: f32,
+    pub kick_radius: f32,
+    pub remaining_amount_of_messages: Option<u16>,
+    pub game_controller_address: Option<SocketAddr>,
+}
+
+impl Default for SimulationConfig {
+    fn default() -> Self {
+        Self {
+            walk_translation_speed: 2.0,
+            walk_rotation_speed: 3.0,
+            walk_with_velocity_scale: 1.0,
+            kick_ball_speed_rumpelstilzchen: 2.0,
+            kick_ball_speed_schlong: 3.0,
+            kick_cooldown: Duration::from_millis(750),
+            ball_friction_per_second: 0.6,
+            ball_visibility_range: 4.0,
+            ball_visibility_angle: std::f32::consts::FRAC_PI_2,
+            robot_radius: 0.25,
+            kick_radius: 0.25,
+            remaining_amount_of_messages: Some(u16::MAX),
+            game_controller_address: None,
+        }
+    }
+}
+
+pub fn default_behavior_parameters() -> Result<BehaviorParameters> {
+    let parameters: serde_json::Value =
+        serde_json::from_str(include_str!("../../../etc/parameters/default.json"))?;
+    let behavior = parameters
+        .get("behavior")
+        .cloned()
+        .ok_or_else(|| eyre!("default parameters do not contain behavior parameters"))?;
+
+    Ok(serde_json::from_value(behavior)?)
+}
