+++ crates/bevyhavior_simulator/src/lib.rs
@@ -1,2 +1,15 @@
+mod auto_referee;
+mod ball;
+mod behavior_runtime;
 pub mod behavior_tree_simulator;
+mod communication;
+mod config;
+mod coordinates;
+mod game_controller;
+mod invariant_checks;
+mod kinematics;
+mod robot;
+mod simulation;
+mod timeline;
 pub mod timeline_viewer;
+mod world_states;
